requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
cffi==1.16.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
fastapi==0.104.1
flatbuffers==23.5.26
fonttools==4.45.1
h11==0.14.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
mediapipe==0.10.8
numpy==1.26.2
opencv-contrib-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
protobuf==3.20.3
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
